Refactorized Course module: - Use STL containers - Fixed signed/unsigned issues
git-svn-id: https://svn.code.sf.net/p/extremetuxracer/code/trunk@285 0420edf4-82e4-42fc-9478-35b55e6d67a3

+++ etr_types.h
@@ -205,14 +205,14 @@ typedef struct {
 	double finish_brake;
 
 	// course and race params 
-	int player_id;
+	size_t player_id;
 	size_t start_player;
 	TCup2* cup;
-	int race_id;
+	size_t race_id;
 	int numraces;
 	int mirror_id;                           
-	int char_id;                           
-	int course_id;
+	size_t char_id;
+	size_t course_id;
 	int location_id;
 	int light_id;
 	int snow_id;
